updated image upload functions. now all working

+++ 1DV508/src/movies/MovieDB.java
@@ -29,8 +29,7 @@ public class MovieDB implements Serializable{
 	private MySQLConnection mysql = new MySQLConnection();
 	
 	private Movie temp = new Movie();
-	private UploadedFile image;
-	private String filename;
+	private UploadedFile image=null;
 	private String searchInput="";
 	private List<Movie> searchResult = this.getMovies();
 	
@@ -204,7 +203,7 @@ public class MovieDB implements Serializable{
 	 * @return return page url
 	 * @throws IOException 
 	 */
-	public String add() throws IOException {
+	public String add(){
 		//	Image upload to the local resources/images/	
 		uploadImage();
 		
@@ -215,7 +214,7 @@ public class MovieDB implements Serializable{
 				stat.setString(1, temp.getTitle());
 				stat.setString(2, temp.getGenre());
 				stat.setString(3, temp.getDescription());
-				stat.setString(4, "images/" + filename);
+				stat.setString(4, temp.getImgPath());
 				stat.setInt(5, temp.getQuantity());
 				stat.setFloat(6, temp.getPrice());
 				stat.executeUpdate();
@@ -233,6 +232,8 @@ public class MovieDB implements Serializable{
 		} catch (ClassNotFoundException e) {
 			e.printStackTrace();
 		}
+		this.setTemp(null);
+		this.setImage(null);
 
 		return "manage_movies";
 	}
@@ -240,19 +241,22 @@ public class MovieDB implements Serializable{
 	
 	public String edit(Movie m) {
 		this.temp = m;
+		this.setImage(null);
 		return "edit_movie";
 	}
 	
-	public String save() throws IOException, InstantiationException, IllegalAccessException, ClassNotFoundException, SQLException{
-		//	The object will find the path to the film's poster.
-		SelectImage image = new SelectImage(this.temp.getId());
-		
-		//	When the path to the image is treated as a file, then you can delete it from the server.
-		File file = new File(image.getImgPath());	
-		file.delete();
-		
-		//	Image upload to the local resources/images/	
-		uploadImage();
+	public String save(){
+		if(!(this.image.getFileName().equals(""))){
+			//	The object will find the path to the film's poster.
+			String tempPath = System.getProperty("user.home") + "/git/1DV508/1DV508/WebContent/resources/" +this.temp.getImgPath();
+			
+			//	When the path to the image is treated as a file, then you can delete it from the server.
+			File file = new File(tempPath);	
+			file.delete();
+			
+			//	Image upload to the local resources/images/	
+			uploadImage();
+		}
 	        
 		try {
 			//	SQL query that updates one movie to the database by id.
@@ -261,7 +265,7 @@ public class MovieDB implements Serializable{
 				stat.setString(1, temp.getTitle());
 				stat.setString(2, temp.getGenre());
 				stat.setString(3, temp.getDescription());
-				stat.setString(4, "images/" + filename);
+				stat.setString(4, temp.getImgPath());
 				stat.setInt(5, temp.getQuantity());
 				stat.setFloat(6, temp.getPrice());
 				stat.setInt(7, temp.getId());
@@ -289,10 +293,10 @@ public class MovieDB implements Serializable{
 		this.temp = x;
 		
 		//	The object will find the path to the film's poster.
-		SelectImage image = new SelectImage(this.temp.getId());
+		String tempPath = System.getProperty("user.home")+"/git/1DV508/1DV508/WebContent/resources/"+this.temp.getImgPath();
 		
 		//	When the path to the image is treated as a file, then you can delete it from the server.
-		File file = new File(image.getImgPath());	
+		File file = new File(tempPath);	
 		file.delete();
 
 		try {
@@ -346,51 +350,23 @@ public class MovieDB implements Serializable{
 	 * 
 	 * @throws IOException
 	 */
-	public void uploadImage() throws IOException {
-		//	Link to the local repository
-		String filePath=System.getProperty("user.home") + "/git/1DV508/1DV508/WebContent/resources/images/";
-				
-		//	Image upload to the local resources/images/			
-        if (null!=image) {
-        	byte[] bytes = image.getContents();
-            filename = FilenameUtils.getName(image.getFileName());
-            BufferedOutputStream stream = new BufferedOutputStream(new FileOutputStream(new File(filePath+filename)));
-            stream.write(bytes);
-            stream.close();
-        }
-	}
-	
-	/**
-	 * Class retrieves image from database by id.
-	 */
-	private class SelectImage {
-		private PreparedStatement stat;
-		private ResultSet rs;
-		private String imgPath;
-		
-		/**
-		 * @param id
-		 * @throws InstantiationException
-		 * @throws IllegalAccessException
-		 * @throws ClassNotFoundException
-		 * @throws SQLException
-		 */
-		public SelectImage(int id) throws InstantiationException, IllegalAccessException, ClassNotFoundException, SQLException {
-			this.stat = mysql.conn().prepareStatement("SELECT * FROM web_shopdb.movies WHERE id = '"+id+"'");
-			stat.execute();
+	public void uploadImage(){
+		try{
+			//	Link to the local repository
+			String filePath=System.getProperty("user.home") + "/git/1DV508/1DV508/WebContent/resources/images/";
+					
+			//	Image upload to the local resources/images/			
+	        if (!(this.image.getFileName().equals(""))) {
+	        	byte[] bytes = image.getContents();
+	            String filename = FilenameUtils.getName(image.getFileName());
+	            BufferedOutputStream stream = new BufferedOutputStream(new FileOutputStream(new File(filePath+filename)));
+	            stream.write(bytes);
+	            stream.close();
+	            this.getTemp().setImgPath("images/" + filename);
+	        }
 			
-			rs = stat.getResultSet();
-			
-			while (rs.next()) {
-				imgPath = rs.getString(5);
-			}
 		}
-		
-		/**
-		 * @return path
-		 */
-		public String getImgPath() {
-			return System.getProperty("user.home") + "/git/1DV508/1DV508/WebContent/resources/" + imgPath;
+		catch(Exception e){	
 		}
-	}
+	}	
 }
